+++ src/bigfloat/bigfloat.rs
@@ -3,7 +3,7 @@
 pub struct BigFloat{
   pub sign: bool,
   pub vals: Vec<i8>,
-  pub decimal: i64,
+  pub decimal: usize,
 }
 /* 
 //Free space for renting
